Remove currency symbol and adjust health metrics
Remove currency symbols from Android and iOS widgets and switch saved-value display to plain numeric strings. Change cigarettes-avoided calculation to use seconds (more precise) and update health improvement to compute 11 minutes per avoided cigarette (returns double). Propagate type change in provider to expose a double for health, add _formatHealth helper to HomeScreen, remove unused currency formatting, and update the UI to use the new formatted health string.

+++ lib/providers/user_profile_provider.dart
@@ -55,5 +55,5 @@ class UserProfileProvider extends ChangeNotifier {
   int getMonthsSinceQuit() => _profile?.getMonthsSinceQuit() ?? 0;
   double getSmokedCigarettesAvoided() =>
       _profile?.getSmokedCigarettesAvoided() ?? 0;
-  int getHealthImprovement() => _profile?.getHealthImprovement() ?? 0;
+  double getHealthImprovement() => _profile?.getHealthImprovement() ?? 0;
 }
